+++ src/api-legacy/v1/borehole/__init__.py
@@ -12,7 +12,6 @@ from bms.v1.borehole.list import ListBorehole
 from bms.v1.borehole.listfiles import ListFilesBorehole
 from bms.v1.borehole.editinglistfiles import EditingListFilesBorehole
 from bms.v1.borehole.ids import BoreholeIds
-# from bms.v1.borehole.export import ExportBorehole
 from bms.v1.borehole.editinglist import ListEditingBorehole
 from bms.v1.borehole.get import GetBorehole
 from bms.v1.borehole.check import CheckBorehole