+++ src/api-legacy/v1/setting/__init__.py
@@ -1,4 +1,3 @@
 
 # Actions
-from bms.v1.setting.download import DownloadHandler
 from bms.v1.setting.patch import PatchSetting