+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Removed _set manually_ from resolving options when deleting a lithology layer.
 - Display lithology layer details in lithology column in editor mode.
 - Immediately delete layers in _casing_ and _backfill_ columns, without showing a delete dialog.
+- Removed export functionality.
 
 ### Fixed